GUI/Graphics: Use int rect for ClipViewScope
This fixes text distortion in ColorPicker.

+++ Essa/GUI/Graphics/ClipViewScope.cpp
@@ -2,7 +2,7 @@
 
 namespace Gfx {
 
-ClipViewScope::ClipViewScope(Gfx::Painter& target, Util::Vector2u host_window_size, Util::Rectf rect, Mode mode)
+ClipViewScope::ClipViewScope(Gfx::Painter& target, Util::Vector2u host_window_size, Util::Recti rect, Mode mode)
     : m_target(target)
     , m_old_projection(target.builder().projection()) {
 
@@ -29,8 +29,8 @@ ClipViewScope::~ClipViewScope() {
     m_target.builder().set_projection(m_old_projection);
 }
 
-llgl::Projection ClipViewScope::create_clip_view(Util::Rectf const& rect, Util::Vector2f offset_position, Util::Vector2u host_window_size) const {
-    return llgl::Projection::ortho({ { offset_position.x(), offset_position.y(), rect.width, rect.height } },
+llgl::Projection ClipViewScope::create_clip_view(Util::Recti const& rect, Util::Vector2f offset_position, Util::Vector2u host_window_size) const {
+    return llgl::Projection::ortho({ { offset_position.x(), offset_position.y(), static_cast<double>(rect.width), static_cast<double>(rect.height) } },
         Util::Recti { static_cast<int>(rect.left), static_cast<int>(host_window_size.y() - rect.top - rect.height), static_cast<int>(rect.width), static_cast<int>(rect.height) });
 }
